pyproject.toml
[tool.poetry]
name = "test-task-didenok-team"
version = "0.1.0"
description = ""
authors = ["peppak <peppius48@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.6"
djangorestframework = "^3.15.2"
cryptography = "^42.0.8"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
